+++ user/src/main/java/pl/gucio/enzo/chronica/user/data/mapper/AccountMapper.java
@@ -1,8 +1,8 @@
-package pl.gucio.enzo.chronica.user.domain.mapper;
+package pl.gucio.enzo.chronica.user.data.mapper;
 
 import org.mapstruct.Mapper;
-import pl.gucio.enzo.chronica.user.domain.entity.AccountEntity;
-import pl.gucio.enzo.chronica.user.domain.model.AccountData;
+import pl.gucio.enzo.chronica.user.data.entity.AccountEntity;
+import pl.gucio.enzo.chronica.user.data.model.AccountData;
 
 @Mapper
 public interface AccountMapper {
