+++ user/src/main/java/pl/gucio/enzo/chronica/user/data/repository/AccountRepository.java
@@ -1,9 +1,9 @@
-package pl.gucio.enzo.chronica.user.domain.repository;
+package pl.gucio.enzo.chronica.user.data.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pl.gucio.enzo.chronica.user.domain.entity.AccountEntity;
+import pl.gucio.enzo.chronica.user.data.entity.AccountEntity;
 
 @Repository
 public interface AccountRepository extends JpaRepository<AccountEntity,Long> {
