+++ user/src/main/java/pl/gucio/enzo/chronica/user/data/repository/PersonRepository.java
@@ -1,8 +1,8 @@
-package pl.gucio.enzo.chronica.user.domain.repository;
+package pl.gucio.enzo.chronica.user.data.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pl.gucio.enzo.chronica.user.domain.entity.PersonEntity;
+import pl.gucio.enzo.chronica.user.data.entity.PersonEntity;
 
 @Repository
 public interface PersonRepository extends JpaRepository<PersonEntity,Long> {
