+++ user/src/main/java/pl/gucio/enzo/chronica/user/data/request/CreateUserRequest.java
@@ -1,0 +1,7 @@
+package pl.gucio.enzo.chronica.user.data.request;
+
+import pl.gucio.enzo.chronica.user.data.model.AccountData;
+import pl.gucio.enzo.chronica.user.data.model.PersonalInformation;
+
+public record CreateUserRequest(AccountData accountData, PersonalInformation personalInformation) {
+}
