--- user/src/main/java/pl/gucio/enzo/chronica/user/domain/request/CreateUserRequest.java
@@ -1,7 +1,0 @@
-package pl.gucio.enzo.chronica.user.domain.request;
-
-import pl.gucio.enzo.chronica.user.domain.model.AccountData;
-import pl.gucio.enzo.chronica.user.domain.model.PersonalInformation;
-
-public record CreateUserRequest(AccountData accountData, PersonalInformation personalInformation) {
-}
